added more tests

+++ tests/EvenireDB.Server.Tests/ServerFixture.cs
@@ -1,8 +1,10 @@
-﻿namespace EvenireDB.Server.Tests
+﻿using Grpc.Net.Client;
+
+namespace EvenireDB.Server.Tests
 {
     public class ServerFixture : IAsyncLifetime
     {
-        private readonly List<WebApplicationFactory<Program>> _instances = new();
+        private readonly List<IDisposable> _instances = new();
 
         public WebApplicationFactory<Program> CreateServer()
         {
@@ -11,16 +13,35 @@
             return application;
         }
 
+        public GrpcChannel CreateGrpcChannel()
+        {
+            var application = CreateServer();
+            
+            var handler = application.Server.CreateHandler();
+            _instances.Add(handler);
+            
+            var channel = GrpcChannel.ForAddress("http://localhost", new GrpcChannelOptions()
+            {
+                HttpHandler = handler,
+            });
+            _instances.Add(channel);
+
+            return channel;
+        }
+
         public Task InitializeAsync()
         => Task.CompletedTask;
 
-        public async Task DisposeAsync()
+        public Task DisposeAsync()
         {
-            Directory.Delete("./data", true);
-
             foreach (var instance in _instances)
-                await instance.DisposeAsync();
+                instance.Dispose();
             _instances.Clear();
+
+            if(Directory.Exists("./data"))
+                Directory.Delete("./data", true);
+
+            return Task.CompletedTask;
         }
     }
 }
